# internal/delivery/http/dto/system.go
package dto

// BreakGlassRequest model for pausing the system or campaign securely
type BreakGlassRequest struct {
	Reason   string `json:"reason,omitempty"`
	Password string `json:"password" validate:"required"`
}